feat: add `raw` module.
`raw` includes the entire public API of `libuiohook`. I decided to not use
`bindgen` to generate these items, because currently it is not feasible to
automatically infer which type should be used for `#define`. That is a major
downside, because that would have been the `number 1` reason to use it. The
library is very small, and doesn't have breaking changes too often so it really
doesn't matter much anyways, if these types are written manually.

Anyways, currently I still don't know if this works at all; I'm still kinda
confused how are you supposed to link a library (or rather, how to deal with
cross-platform .so selection). Once I have that figured out, I will start
porting the examples provided by the project to check if everything works
correctly.